# code/frontend/src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/searches": {
      "post": {
        "tags": [
          "search-controller"
        ],
        "operationId": "createSearch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RealEstateResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/realestates/{realestateid}/offers": {
      "get": {
        "tags": [
          "offer-controller"
        ],
        "operationId": "getPagedRealEstateOffers",
        "parameters": [
          {
            "name": "realestateid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfferResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "offer-controller"
        ],
        "operationId": "createOffer",
        "parameters": [
          {
            "name": "realestateid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/notificationcategories/{notificationcategoryname}/notifications": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "getNotifications",
        "parameters": [
          {
            "name": "notificationcategoryname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Notification"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "createNotification",
        "parameters": [
          {
            "name": "notificationcategoryname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/estateagents": {
      "post": {
        "tags": [
          "estate-agent-controller"
        ],
        "operationId": "registerEstateAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StafferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstateAgentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admins": {
      "post": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "registerAdmin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StafferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/realestates/{realestateid}/offers/{offerid}": {
      "get": {
        "tags": [
          "offer-controller"
        ],
        "operationId": "getOfferById",
        "parameters": [
          {
            "name": "realestateid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offerid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "offer-controller"
        ],
        "operationId": "updateOfferStatus",
        "parameters": [
          {
            "name": "realestateid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offerid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/notificationcategories/{notificationcategoryname}/notifications/{notificationid}": {
      "get": {
        "tags": [
          "notification-controller"
        ],
        "operationId": "getNotificationById",
        "parameters": [
          {
            "name": "notificationcategoryname",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "notificationid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admins/{adminid}": {
      "get": {
        "tags": [
          "admin-controller"
        ],
        "operationId": "getAdminById",
        "parameters": [
          {
            "name": "adminid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "RealEstateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "category": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "description": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "estateAgentEmail": {
            "type": "string"
          },
          "detailId": {
            "type": "integer",
            "format": "int64"
          },
          "cadastralDataId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CadastralFilter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "priceRange": {
            "$ref": "#/components/schemas/PriceRange"
          },
          "squareMetersRange": {
            "$ref": "#/components/schemas/SquareMetersRange"
          },
          "energyClassRange": {
            "$ref": "#/components/schemas/EnergyClassRange"
          },
          "roomsRange": {
            "$ref": "#/components/schemas/RoomsRange"
          },
          "floorRange": {
            "$ref": "#/components/schemas/FloorRange"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "search": {
            "$ref": "#/components/schemas/Search"
          }
        }
      },
      "Detail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "geographicalPosition": {
            "$ref": "#/components/schemas/GeographicalPosition"
          },
          "utility": {
            "$ref": "#/components/schemas/Utility"
          }
        }
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "message": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "notificationCategory": {
            "$ref": "#/components/schemas/NotificationCategory"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "NotificationCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "enum": [
              "NEW_PROPERTIES",
              "PROMOTIONAL",
              "VISIT",
              "OFFER"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          }
        }
      },
...
      "Proposal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "proposalCategory": {
            "type": "string",
            "enum": [
              "OFFER",
              "COUNTER_OFFER",
              "VISIT"
            ]
          },
          "proposalStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ]
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "realEstate": {
            "$ref": "#/components/schemas/RealEstate"
          }
        }
      },
...
      "Search": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "category": {
            "type": "string",
            "enum": [
              "SALE",
              "RENT"
            ]
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "searchRealEstates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchRealEstate"
            }
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "detail": {
            "$ref": "#/components/schemas/Detail"
          },
          "cadastralFilter": {
            "$ref": "#/components/schemas/CadastralFilter"
          }
        }
      },
      "SearchRealEstate": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/SearchRealEstateKey"
          },
          "realEstate": {
            "$ref": "#/components/schemas/RealEstate"
          },
          "search": {
            "$ref": "#/components/schemas/Search"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SearchRealEstateKey": {
        "type": "object",
        "properties": {
          "realEstateId": {
            "type": "integer",
            "format": "int64"
          },
          "searchId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "searches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Search"
            }
          },
          "proposals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Proposal"
            }
          }
        }
      },
...
      "Utility": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "hasAirConditioning": {
            "type": "boolean"
          },
          "hasDoorman": {
            "type": "boolean"
          },
          "hasElevator": {
            "type": "boolean"
          },
          "nearSchool": {
            "type": "boolean"
          },
          "nearPublicTransport": {
            "type": "boolean"
          },
          "nearPark": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "detail": {
            "$ref": "#/components/schemas/Detail"
          }
        }
      },
      "SearchRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "minLength": 1
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "cadastralFilterId": {
            "type": "integer",
            "format": "int64"
          },
          "detailId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "category",
          "page",
          "size"
        ]
      },
...
      "OfferRequest": {
        "type": "object",
        "properties": {
          "category": {
            "type": "string",
            "minLength": 1
          },
          "status": {
            "type": "string",
            "minLength": 1
          },
          "amount": {
            "type": "number"
          },
          "counteredOfferId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "amount",
          "category",
          "status"
        ]
      },
      "OfferResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "userEmail": {
            "type": "string"
          },
          "realEstateId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          },
          "counteredOfferId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "NotificationRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1
          },
          "userEmail": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "message",
          "userEmail"
        ]
      },
      "NotificationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          },
          "message": {
            "type": "string"
          },
          "notificationCategoryId": {
            "type": "integer",
            "format": "int64"
          },
          "userEmail": {
            "type": "string"
          }
        }
      },
      "StafferRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email"
        ]
      },
      "EstateAgentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "adminEmail": {
            "type": "string"
          },
          "createdRealEstates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RealEstateResponse"
            }
          }
        }
      },
...
      "AdminResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "adminEmail": {
            "type": "string"
          },
          "createdStaffers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StafferResponse"
            }
          }
        }
      },
      "StafferResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "adminEmail": {
            "type": "string"
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageOfferResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OfferResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      }
    }
  }
}